requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
ollama==0.5.1
mysql-connector-python==9.3.0
jinja2==3.1.6
requests==2.32.4
mypy==1.16.1
pre-commit==4.2.0
pyOpenSSL==25.1.0
pydantic==2.11.7
python-dotenv==1.1.0
pvporcupine==3.0.5
pyaudio==0.2.14
pywifi==1.1.12
packaging==25.0
beautifulsoup4==4.13.4